requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
Pillow==9.5.0
colorthief==0.2.1
webcolors==1.12.0
tensorflow==2.15.0  # Contains built-in keras
opencv-python==4.7.0.72
numpy==1.24.3
selenium==4.10.0
h5py==3.10.0
werkzeug==2.0.3
